Only initialise PTM for nodes (#1152)
* Only initialise the PTM connection when running a node instance, instead of a subcommand.

* Move PTM init into Quorum section
